+++ src/app/components/provider-place/drop.js
@@ -29,6 +29,7 @@ export default Component.extend(I18n, {
   guiUtils: service(),
   router: service(),
   providerResources: service(),
+  globalClipboard: service(),
 
   /**
    * @virtual
@@ -170,12 +171,8 @@ export default Component.extend(I18n, {
   },
 
   actions: {
-    copySuccess() {
-      this.globalNotify.info(this.t('hostnameCopySuccess'));
-    },
-
-    copyError() {
-      this.globalNotify.info(this.t('hostnameCopyError'));
+    copyHostToClipboard() {
+      this.globalClipboard.copy(this.provider.domain);
     },
   },
 });