+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Latest changes
 
+* VFS-13461 Added provider version on providers page
 * VFS-13513 Fixed white files option without privileges in space sidebar
 * VFS-13395 Added searching by id in space sidebar
 * VFS-13511 Fixed docs and API links for new homepage